requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.15.0
dash>=2.10.0

# Time series analysis
statsmodels>=0.14.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Deep learning for forecasting
tensorflow>=2.13.0
keras>=2.13.0

# Data collection and APIs
requests>=2.31.0
aiohttp>=3.8.0
python-dotenv>=1.0.0

# Geographic data and mapping
folium>=0.14.0
geopandas>=0.13.0
cartopy>=0.21.0
basemap>=1.3.0

# Configuration and data handling
PyYAML>=6.0
xarray>=2023.6.0
netcdf4>=1.6.0
h5py>=3.9.0

# Time zone handling
pytz>=2023.3
tzdata>=2023.3

# Progress bars and utilities
tqdm>=4.65.0
click>=8.1.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Jupyter notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
ipywidgets>=8.0.0

# Development tools
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0

# Data validation
pydantic>=2.0.0
cerberus>=1.3.0

# Performance optimization
numba>=0.57.0
dask>=2023.7.0

# Climate-specific libraries
xclim>=0.44.0
cf-xarray>=0.8.0

# Statistical testing
pingouin>=0.5.0
arch>=5.5.0  # For time series statistical tests